+++ packages/payment-integration-api/src/payment-provider-customer/payment-provider-customer.ts
@@ -1,0 +1,10 @@
+import { AddressRequestBody } from '../address';
+import { CardInstrument } from '../payment';
+
+export type PaymentProviderCustomer = BraintreeAcceleratedCheckoutCustomer;
+
+export interface BraintreeAcceleratedCheckoutCustomer {
+    authenticationState?: string;
+    addresses?: AddressRequestBody[];
+    instruments?: CardInstrument[];
+}